move errors to separate module
