+++ src/lib.rs
@@ -27,6 +27,7 @@
 //! # }
 //! ```
 
+pub mod errors;
 pub mod http;
 pub mod jsonrpc;
